修改更改

+++ model/LoginUser.go
@@ -1,7 +1,7 @@
 package model
 
 type LoginUser struct {
-	ID       int64  `json:"id" gorm:"primary_key" form:"id"`
+	ID       uint   `json:"id" gorm:"primary_key" form:"id"`
 	UserName string `json:"username" form:"username"`
 	Password string `json:"password" form:"password"`
 }
